Partially de-duplicate flood_model/*/*_parameters_generator.py
