+++ src/eddie_floodresilience/tables.py
@@ -84,9 +84,9 @@ class RiverNetwork(Base):
     geometry = Column(Geometry("POLYGON", srid=2193))
 
 
-class BGFloodModelOutput(Base):
+class FloodModelOutput(Base):
     """
-    Class representing the 'bg_flood_model_output' table.
+    Class representing the 'flood_model_output' table.
 
     Attributes
     ----------
@@ -104,10 +104,10 @@ class BGFloodModelOutput(Base):
         Geometric representation of the catchment area coverage.
     """  # pylint: disable=too-few-public-methods
 
-    __tablename__ = "bg_flood_model_output"
+    __tablename__ = "flood_model_output"
     unique_id = Column(Integer, primary_key=True, autoincrement=True)
-    file_name = Column(String, comment="name of the flood model output file")
-    file_path = Column(String, comment="path to the flood model output file")
+    file_name = Column(String, comment="name of the flood model output file", nullable=True)
+    file_path = Column(String, comment="path to the flood model output file", nullable=True)
     created_at = Column(DateTime(timezone=True), default=datetime.now(timezone.utc), comment="output created datetime")
     geometry = Column(Geometry("POLYGON", srid=2193))
 
@@ -128,7 +128,7 @@ class BuildingFloodStatus(Base):
     is_flooded : bool
         If the building is flooded or not
     flood_model_id: int.
-        Foreign key mathing the unique_id from bg_flood_model_output table
+        Foreign key mathing the unique_id from flood_model_output table
     """  # pylint: disable=too-few-public-methods
 
     __tablename__ = "building_flood_status"
